Fix CentOS 7 image build on aarch64 (#46)
CentOS 7 is an important OS for our CI/CD pipeline, probably with the oldest version of glibc of any supported Linux distribution as of 10/2022. By building software on CentOS 7, we can produce packages that work on most glibc-supported Linux distributions. This patch enables building a CentOS 7 Docker image for aarch64.

Software version changes:
- Using Golang version 1.19.2 on aarch64 CentOS 7, by installing from a tarball and not from a repository. We should probably switch to this way of installing Golang across the board, for all other Linux versions as well.
- Updating the version of Python installed from source from 3.8.2 to 3.8.14.

Also generating more locales on RHEL-based systems. Previously, even the en_US.UTF-8 locale was not being generated on CentOS 7. The list of locales was provided by @andrei-mart based on what is needed by Postgres for internationalization.